comida_al_paso_project/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
psycopg2-binary
gunicorn
whitenoise
